Moves model definition code into its own file
