+++ app.js
@@ -1,209 +1,17 @@
 var express = require('express'),
     mongoose = require('mongoose'),
-    Schema = mongoose.Schema,
     mongooseAuth = require('mongoose-auth'),
     conf = require('./conf'),
     markdown = require('markdown').markdown,
     _ = require('underscore'),
+    models = require('./models'),
 
-    Project,
-    PersonSchema,
-    Person,
-    JobPostSchema,
-    JobPost,
-    JobRequestSchema,
-    JobRequest,
     app,
-    handleError;
-
-// Model config
-Project = new Schema({
-  name        : String, 
-  project_url : String,
-  description : String
-});
-
-PersonSchema = new Schema({
-  name          : String,
-  email         : String,
-  gravatar      : String, // MD5 hash based on email
-  irc           : String,
-  twitter_nick  : String,
-  github_nick   : String,  
-  bio           : String,
-  url_slug      : String,
-  languages     : [String],
-  projects      : [Project],
-  active        : Boolean
-});
-
-PersonSchema.plugin(mongooseAuth, {
-  everymodule: {
-    everyauth: {
-      User: function () {
-        return Person;
-      }
-    }
-  }, 
-  twitter: {
-    everyauth: {
-      myHostname: 'http://localhost:3000',
-      consumerKey: conf.twit.consumerKey,
-      consumerSecret: conf.twit.consumerSecret,
-      redirectPath: '/',
-      findOrCreateUser: function (session, accessTok, accessTokSecret, twitterUser) {
-        var promise = this.Promise(),
-            self = this;
-
-        Person.findOne(
-          { 
-            $or: 
-            [
-              { 'name': twitterUser.name },
-              { 'twitter_nick': twitterUser.screen_name }
-            ]
-          }, 
-          function (err, person) {
-          if (err) return promise.fail(err);
-
-          if (person) {
-            return promise.fulfill(person);
-          } else {
-            person = new Person();
-            person.name = twitterUser.name;
-            person.twitter_nick = twitterUser.screen_name;
-            person.bio = twitterUser.description;
-
-            // Ok, so we're defaulting to all active accounts
-            // for now. In the future, it may be useful to 
-            // prevent anybody from signing up willy nilly
-            // When that day comes, we can make sure new accounts
-            // are inactive upon creation so they can be approved
-            // later            
-            person.active = true;
-
-            person.save(function (e, p) {
-              if (e) return promise.fail(e);
-              return promise.fulfill(p);
-            });
-          }
-        });
-
-        return promise;
-      }
-    }          
-  },
-  password: {
-    loginWith: 'email',
-    extraParams: { name: String },
-    everyauth: {
-      getLoginPath: '/login',
-      postLoginPath: '/login',
-      loginView: 'sessions/login.jade',
-      getRegisterPath: '/register',
-      postRegisterPath: '/register',
-      registerView: 'sessions/register.jade',
-      loginSuccessRedirect: '/',
-      registerSuccessRedirect: '/'
-    }       
-  },
-  github: {
-    everyauth: {
-      myHostname: 'http://localhost:3000',
-      appId: conf.github.appId,
-      appSecret: conf.github.appSecret,
-      redirectPath: '/',
-      findOrCreateUser: function (session, accessTok, accessTokExtra, githubUser) {
-        var promise = this.Promise(),
-            self = this;
-
-        Person.findOne(
-          { $or: 
-            [ 
-              { 'name': githubUser.name },
-              { 'github_nick': githubUser.login }, 
-              { 'email': githubUser.email }
-            ]
-          }, 
-          function (err, person) {
-            if (err) return promise.fail(err);
-
-            if (person) {
-              // Found an existing person
-              return promise.fulfill(person);
-            } else {
-              // Person doesn't already exist with that github info
-              person = new Person();
-              person.name = githubUser.name;
-              person.email = githubUser.email;
-              person.github_nick = githubUser.login;
-
-              // Ok, so we're defaulting to all active accounts
-              // for now. In the future, it may be useful to 
-              // prevent anybody from signing up willy nilly
-              // When that day comes, we can make sure new accounts
-              // are inactive upon creation so they can be approved
-              // later
-              person.active = true;
-
-              person.save(function (e, p) {
-                if (e) return promise.fail(e);
-                return promise.fulfill(p);
-              });
-            }
-          }
-        );
-
-        return promise;
-      }
-    }         
-  }
-});
-
-PersonSchema.pre('save', function (next) {
-  /*
-  * Generate an MD5 hash of the supplied email
-  * and save that as the gravatar string before saving
-  */
-  if (this.email) {
-    this.gravatar = require('./MD5').toMD5(this.email);
-  }
-
-  /*
-  * Remove spaces and weirdo characters to make an addressable
-  * slug for this person. Hope people don't have the same names...
-  */
-  this.url_slug = this.name.toLowerCase().replace(/\s/g, '-').replace(/[^a-z0-9\-]/g, '');
-
-  next();
-});
-
-JobPostSchema = new Schema({
-  headline      : String,
-  company_name  : String,
-  description   : String,
-  category      : {type: String, enum: ['ft', 'pt', 'fl', 'ct']}, /* full-time, part-time, freelance, contract */
-  info_url      : String,
-  contact_email : String,
-  technologies  : [String],
-  date_created  : Date
-});
-
-JobPostSchema.pre('save', function (next) {
-  this.date_created = this.date_created || new Date();
-  next();
-});
-
-JobRequestSchema = new Schema({
-  headline      : String,
-  category      : {type: String, enum: ['ft', 'pt', 'fl', 'ct']}, /* full-time, part-time, freelance, contract */    
-  technologies  : [String],
-  date_created  : Date
-});
-
-JobRequestSchema.pre('save', function (next) {
-  this.date_created = this.date_created || new Date();
-});
+    handleError,
+    Person, 
+    Project,
+    JobPost, 
+    JobRequest;
 
 // Configuration
 app = express.createServer(
@@ -212,8 +20,7 @@ app = express.createServer(
   express.favicon(__dirname + '/public/favicon.ico'),
   express.static(__dirname + '/public'),
   express.cookieParser(),
-  express.session({ secret: 'a1b2c3d4' }),
-  mongooseAuth.middleware()
+  express.session({ secret: 'a1b2c3d4' })
 );
 
 app.configure(function () {
@@ -241,20 +48,24 @@ app.configure('production', function() {
   app.set('db-uri', 'mongodb://localhost/bandc206-production');
 });
 
-console.log(app.set('db-uri'));
 mongoose.connect(app.set('db-uri'));
+models.defineModels(
+  mongoose, 
+  mongooseAuth, 
+  Project,
+  Person,
+  JobPost,
+  JobRequest,
+  function () {
+    app.Person = Person = mongoose.model('Person');
+    app.JobPost = JobPost = mongoose.model('JobPost');
+    app.JobRequest = JobRequest = mongoose.model('JobRequest');
+});
 
-mongoose.model('Person', PersonSchema);
-mongoose.model('JobPost', JobPostSchema);
-mongoose.model('JobRequest', JobRequestSchema);
-
-Person = mongoose.model('Person');
-JobPost = mongoose.model('JobPost');
-JobRequest = mongoose.model('JobRequest');
-
-// Bootstrap
-// require('./bootstrap').bootstrap(Person);
-
+app.configure(function () {
+  app.use(mongooseAuth.middleware());
+});
+  
 var doAuth = function (req, res, next) {
   if (req.loggedIn) {
     next();
